Ensure view exists before updating
When updating a view, we drop it first. This will cause an error if the
view did not previously exist, which is what we want.

We then create the view using `CREATE VIEW` rather than `CREATE OR
UPDATE`.

In addition to being safer than blindly `CREATE OR UPDATE`ing, this also
works across every relational database I know of that supports views,
where `CREATE OR UPDATE` does not.