Use keyword arguments for version specifier
This provides a more readable API:

* `create_view :searches, version: 15`
* `update_view :searches, version: 2`

It also has the added benefit of being more easily extensible as we move
forward (`create_view :searches, materialized: true`) without having to
do craziness to inspec the final argument and do something based on its
type.

This requires upping the required ruby version to 2.x. I decided to keep
compatibility with 2.0.x by not relying on required keyword arguments.
There are a number of people still running 2.0.x due to garbage
collection issues in 2.1. This means I had to add an explicit check to
`update_view` that the required `version` argument was passed.