Make drop_view reversible
`drop_view` is reversible if it is called with a `revert_to_version`
argument. This is the version that `create_view` will use.

This change removes the railtie integration spec in favor of a specific
test of the revert functionality that also happens to test exactly what
the railtie spec was testing.