Make update_view reversible
For `update_view` to be reversible, it must be passed
`revert_to_version`.

This change extracts the nested `ScenicArguments` to a proper object
(`StatementArguments`) and *correctly* inverts the version argument
while maintaining any other arguments that might have been passed.

This also corrects the inversion of drop_view to create the version
specified in `revert_to_version`.