scenic 0.1.0
Scenic makes it easier to work with Postgres views in Rails.

It introduces view methods to ActiveRecord::Migration and allows views
to be dumped to db/schema.rb.

It provides generators for models, view definitions, and migrations.

It is build around a basic versioning system for view definition files.

In short, go add a view to your app.