Drop existing view before attempting to create
Reloading a schema from scratch (on development or test) is causing an
error when ActiveRecord attempts to recreate views that are already
present in the database (in my case, it was `pg_stat_statements`). This
change uses `CREATE VIEW`'s `OR REPLACE` argument to replace the view
if one already exists.

Fixes #55.