Don't dump views for Postgres extensions
If a database view has the same name as a database extension, we want to
ignore that view and not dump it with the rest of the schema. (The
extension will already be handled through the schema's
`enable_extension` statement.) This approach will handle system views
from extensions like PostGIS whose views do not begin with `pg_`.

WIP