Don't dump views for extensions
If a database view has the same name as a database extension, we want to
ignore that view and not dump it with the rest of the schema. (The
extension will already be handled through the schema's
`enable_extension` statement.) This approach will handle system views
from extensions like PostGIS whose views do not begin with `pg_`.

For database types that don't support extensions, ActiveRecord returns
an empty array, so this change will not affect those database adapters
should they be added in the future.