Don't support updating materialized views
Materialized views are frequently used because they can be indexed. Our
strategy for updating a view is to drop it and recreate it. This would
cause us to lose all previously existing indexes.

Ideally we would support recreating the indexes but that can be
exceedingly difficult. We can get the list of indexes on a table from
the database but we'd have to filter out any indexes where columns may
have been dropped.