Document public API
* Added documentation where it was missing.
* Updated documentation to use yard formatting throughout.
* Marked internals as private (`SchemaDumper`, `CommandRecorder`, etc)

I did opt to document `Adapters::Postgres` and `View`. They are not
objects we expect end users to interact with, but the documentation will
be useful for anyone who wants to create an adapter.

Those objects are tagged with `@api extension` so we could filter them
from the default documentation with and update to the `.yardopts` file
if needed.