Adding support to refresh materialized views concurrently
This PR adds support for refreshing materialized views concurrently. I've added the necessary code to the Postgres adapter, and also updated the model generator to include a new `refresh_concurrently` method.

Since the ability to refresh a materialized view concurrently is dependent on a unique index without a `where` clause on that view, it is possible that calling the method will raise an exception. I thought about rescuing that exception and logging a message for the user to let them know that a concurrent refresh wasn't possible, but I felt that in this case an exception should be raised so the user doesn't think the refresh succeeded if they're not looking at their logs or for some reason doesn't see the console output. I've updated the README so there is documentation that the exception will be raised in those cases, so hopefully users won't be surprised.

There was also a small typo in `CONTRIBUTING.md`, which I've fixed here as well. It was so small I didn't think it deserved its own PR.