Add support for updating materialized views
Updating materialized views used to be an error because providing a new
definition for a materialized view requires dropping and recreating it.
This will cause any indexes on the old materialized view to be dropped.
For this reason, we suggested that users manually `drop_view`,
`create_view` and `add_index`. This is a frustrating user experience.

This changes view updates to record any indexes that were in place
before the `drop_view` and then re-create those indexes once the new
materialized view is in place. Any indexes that are now invalid
(most-likely because the column no longer exists) will be dropped. This
behavior is similar to what happens when you alter the schema of a
table, so I do not believe it to be surprising.

We log the results of each index creation for the users information in
the migration output. For example:

```
== 20151228201117 UpdateSearchesToVersion2: migrating =========================
-- update_view(:searches, {:version=>2, :revert_to_version=>1, :materialized=>true})
   -> index 'searches_test_1' on 'searches' has been recreated
   -> index 'searches_test_2' on 'searches' is no longer valid and has been dropped.
   -> 0.0148s
== 20151228201117 UpdateSearchesToVersion2: migrated (0.0149s) ================
```

Closes #98