Add better error messages for materialized views
Previous to this change, attempts to use materialized view functionality
on versions of Postgres that did not support them would result in
ActiveRecord raising a Postgres syntax error. This reqires a small jump
on the part of the user to go from that syntax error to "Oh, my version
doesn't support materialized views."

This change does some version detecting and pre-emptively raises more
descriptive errors for these cases, so that jump is no longer required.