Fix failing test in Rails 4.1
When selecting the first row from a materialized view in Rails 4.1,
ActiveRecord was automatically adding an `ORDER BY id` clause even
though the table does not have an Id. We can't control the behavior of
`first` in ActiveRecord, but we can use `take` instead which does not
have this behavior.