Support versions of Postgres older than 9.3
Postgres pre 9.3 does not support materialized views. Scenic can work
just fine for users of those versions, but it does not because the
`views` method on the adapter assumes that we can query `pg_matviews`
which does not exist on those older versions.

In this change, we introduce the concept of a decorated connection, with
the undecorated connection being injected into the adapter. This gives
us a place to decorate the connection with methods about its
capabilities. Some of these methods are defined in newer versions of
rails or are public in newer versions while being private in older
versions. We try to use the Rails versions of these methods where
possible before using our own implementation.

As part of this change, the code that looks up the views and returns
`Scenic::View` objects was extracted to its own class. This gave it some
space to grow and made for a more readable implemnentation than
otherwise possible.