Make API docs for Postgres adapter public
While we don't expect people to call these methods directly, the
documentation they offer is likely to answer questions on how Scenic
interfaces directly with Postgres.

Additionally, this is the only place `#refresh_materialized_view` is
documented. This was already a public method (we put it in generators),
so documenting it seems wise.

The method names in use in the adapter are often duplicates of the
method names exposed in migrations. For that reason, I specifically
linked from the adapter method to its corresponding method in the schema
statements.