Adding support for views in schemas other than public
I ran into an issue where I wanted to create a view in a schema on Postgres other than the `public` schema, and the generator gave me files that weren't quite what I needed. It wasn't all that tricky to fix them, but I figured it would be easy enough to add in support for them for good measure.

Basically the only thing we had to do was to change the generator, and specifically we needed to override the Rails default method for `plural_file_name` so that it removed the `.` needed in the table definition when the schema was included. I also had to change the table name in the actual migration from a symbol to a string to support having a `.` in the name.