Include materialized view indexes in schema dumps
I forgot that we also need to instruct the schema dumper to dump any
indexes associated to our materialized views. Prior to this change,
materialized view indexes would not survive a round trip to and from
`schema.rb`.

The change was relatively simple because we can rely on the previously
existing `indexes` method provided by Rails to do the work. I included
some formatting changes to make the dumped file consistent with what
rails outputs.

Rails outputs everything in schema with a trailing newline with the
exception of foreign keys, which do not have a final trailing newline
for whatever reason. This means we need to insert a newline before our
view-related dumping. Unfortunately, this will lead to:

1. A double newline at the start if the schema has no foreign keys.
2. An extra newline at the bottom (before `end`).

There's no good way to fix this so we'll live with it.