Fix bug in matview index reapplication
The use of `quote_table_name` was causing the index reapplication code
to not match any indexes on the supplied view name. This means the index
reapplication feature was not working. We had insufficient test coverage
to catch this. I have added test coverage for this as part of the fix.