Add update_view_replace migration statement
Despite the fact that drop_view, create_view is usually the better
choice there are some situations where CREATE OR REPLACE VIEW can be
better. An example is when you have a complicated dependency tree
between your views or when the view is depended upon by a slow
materialized view.

This commit adds a update_view_replace migration statement (and
associated code) that uses CREATE OR REPLACE VIEW instead of
dropping and creating the view.