Add view namespaces to SchemaDumper
Add non-public namespaces to view names when dumping to db/schema.rb.
This is useful when views live in another schema than `public`. Before
this change the dumper would dump `thoughtbot.scenic` as `scenic`. In
our case this would collide with our `scenic` view that's living in the
`public` namespace.