Adds #replace_materialized_view
Now, when you specify replace_view on a migration (on a materialized view), It will first create the new materialized view (under a temporary name),
including the indexes defined in that version. Then it will drop the previous existing materialized view, and rename the new materiliazed view to its original name. Modify spec.