Add refreshing materialized views with dependencies
We had issues with refreshing complex chains of materialized views that
depended on one another. For example, we had the following type of thing
happening.

(imagine all the relations below are materialized views)

A depends on B
B depends on C and D
C depends on D
D depends only on non-materialized views

Now if you wanted to refresh A, and you wanted it to be as fresh as possible,
first you would need to refresh the materialized views on which it depends,
followed by A itself. This became frequent enough, and changed with enough
frequency, to make keeping refresh methods for each individual materialized
view unreasonable.

I realized we could use the information kept in `pg_depend` for this purpose,
and also thought that I'm most likely not the first person to have this
problem, so I'm contribuing it back here. Also, I can see further improvements
to the migrations being built off of this usage of the dependency graph stored
in Postgres, so I thought it might help!