Update bin/setup and acceptance_helper.rb
Some folks were encoutering issues with the additional security layer for
databases that came in Rails 5, which was preventing the dropping of the test
database. This commit adds in some conditional stuff to do that additional
setup for Rails 5.

@derekprior - for some reason I wasn't able to see this fail on my machine, but
I'm curious if it does the job for you. I was able to do the five things you
listed in the issue without problem, so hopefully it'll work!

Resolves #177