Address further comments in cascading materialized view refresh (#180)
Derek had a few other comments about this feature that didn't get
addressed before it was merged, so this PR is to address those issues. It aims
to make the process of parsing the dependency information from Postgres a
little easier to understand, addresses some style issues, and also adds
documentation.