Add Rails 5.1 and edge Rails to build matrix
We previously had an entry for edge rails but it was configured
incorrectly in the travis matrix. It was listed as an exclude rather
than an allowed failure.

This also removes the dedicated appraisal added for a single factory
girl related test and that test. I think the test was useful during
development, but not on an ongoing basis.