Add `no_data` option to materialized views
By default, creating a materialized view causes the associated query to
be run immediately. The `create view` statement does not complete until
the view is populated. While this is a good default, it may be
particularly expensive and potentially not desired during deploys.
Specifying `no_data: true`, allows the view to be created without a
refresh. It will not be usable until it is manually refreshed.