Gracefully handle case where view has no dependencies
We were previously assuming that anytime a user passed `cascade: true`
to `refresh_materialized_view`, that the view they were attempting to
refresh actually had dependencies. Turns out, we would end up with a
weird exception in this case.

I've changed it so the `cascade: true` part of that operation is
essentially a NOOP now, not raising any exception. I figured that it
would still accomplish what the user set out to do (refresh their
materialized view) without requiring any additional understanding or
changing of code.