Handle semicolon when using `no_data: true`
Some folks add `;` to the end of their view definitions to terminate the
statement. When this is combined with `no_data: true`, an exception was
being raised because `WITH NO DATA` was being interprested as a
separate SQL statement.

To fix this, we remove the trailing semicolon from the SQL definition
before adding `WITH NO DATA`.

Fixes #272