Rename NEWS.md to CHANGELOG.md
Semantically, I agree that the changelog for a release is the list
of every commit and that what we typically refer to as a changelog these
days is really a list of newsworthy changes. However, it seems to me
that battle is lost and that people expect this file to be named
`CHANGELOG.md`. So here it is.