Fix postgres configuration in CI workflow
I'm not sure when or why this stopped working, but it seems we now need
to explicitly set the `ENV` to be used in the postgres service. It used
to inherit (or perhaps default) the correct values.

While I was in here, I also took the version specifier off the Postgres
image so we always test against the latest release available.