Use pg 1.1 or newer in tests
Edge rails now requires pg 1.1, but can't do so as a dependency in the
Gemfile because bundler doesn't have optional dependencies. The result
is that bundler is perfectly happy to install an older version of the pg
gem, but it blows up once we start the Rails environment.

There were a few options:

1. Update the gem dependency in the `gemspec` to 1.1+ - This seemed
extreme to force on folks.

2. Use the newer gem only in the edge rails branch of the gemfile -
Sure, this seems reasonable.

3. Use the newer gem in all our tests, even though older versions would
work.

I opted for the third option just because I think people who are
updating Scenic are also likely to be updating other gems, like pg.