Add Security Policy
GitHub recently added support for special handling of security policy
files. To date, we've not received any security vulnerability reports,
but it still seemed like a good idea to have this in place.

I am not an expert, but I think I captured what we'd actually want to do
here.