Allow view creation using raw SQL input
`create_view` helps the user by wrapping their
`SELECT` statement into a `CREATE VIEW` statement.
This is great in most cases but it breaks the
ability to build complex views that depends on
other views.

This can already be done by individually migrating
several views. However I find that if the goal is
to create a single complex view it's better to
contain all the necessary subviews inside the same
SQL file.