Fix issue with keyword arguments under Ruby 3
Ruby 3 changed the handling of keyword arguments in a way that broke our
liberal handling of arguments in the command recorder. Using
`ruby2_keywords` helps us fix this while maintaining compatibility with
older (pre 2.7) versions of Ruby.

In the next major version of Scenic, we should drop support for rubies
older than either 2.7 or 3.0 to avoid carrying this fix forward.

Fixes #331. For more context on why this is the solution, see:

* https://eregon.me/blog/2021/02/13/correct-delegation-in-ruby-2-27-3.html
* https://github.com/rails/rails/issues/42653