Add devcontainer configuration for Codespaces
This configuration allows Scenic development to be done via Codespaces
for anyone with access to that feature on GitHub. Creating a Codespace
in this project will drop you into an environment ready to run `rake`
and get a fully passing test suite.

I personally do this using the `gh` cli, connecting via SSH, but one
could also connect via VSCode or the web.

Either way, I'm in a fully working environment within 1-2 minutes, which
I find helpful when I've stepped away from development work for a bit
and can't be bothered to figure out my local environment again.

I believe this same configuration should be usable locally for folks who
want to use Docker (see #319), but that is less interesting to me
personally.