Exclude pg_stat_statements_info
Postgres 14 [introduced] an internal view pg_stat_statements_info, which
we need to exclude from our dumper because including it causes loading
the schema to fail:

    PG::DuplicateTable: ERROR:  relation "pg_stat_statements_info" already exists

To test this, enable pg_stat_statements in the dummy database. This is a
reasonable addition as the majority of rails apps are using this
extension and it's cleaner and less error-prone than adding a fake view
of the same name in a specific test for this case.

To support this view, we're currently specifically excluding it in the
query used in the Postgres adapter to load views. Rather then excluding
it by name, we could do an extension of our current exclusion for
extension-owned views, `AND c.relname NOT IN (SELECT extname FROM
pg_extension)` by excluding any views with an extension name included
rather than exact matches. I'm not convinced that this is more robust
than excluding exact names has been shown to be by this bug.

Fixes #348