Rename `--replace_view` flag to `--replace`
This flag was introduced in #368 but hasn't been in a release yet. I was
writing up the next release's changelog and I wasn't thrilled with
having an underscore in a CLI flag.

There's something to be said for `--replace_view` matching the schema
method that the generated method invokes but I think `--replace` is
still better.