Remove rails master from our test matrix
We're not an active enough project to keep up with master as part of a
required build task. I don't want to turn off any contributors who see a
red build against an unreleased version of Rails.

I think we could rework these as a separate workflow that reports status
appropriately but is not required, but I'll leave that for another day.