Only refresh concurrently if view is populated
This implementation will still error if you ask for a concurrent refresh
of a non-populated view if your database does not support concurrent
refreshes. I'm happy with this because the user has asked for a thing
that we know cannot succeed, even under the right circumstances, and we
should tell them that., and we should tell them that.

Co-authored-by: Derek Prior <derekprior@gmail.com>
Co-authored-by: Caleb Hearth <caleb@calebhearth.com>