Update CI Versions
It's time for everyone's favorite game: bump the versions! Ruby 3.4 is
out, and Rails 8 has been shipped for some time. We try to keep our CI
suite testing currently-supported versions of each, but we don't want a
giant matrix that requires many customizations to run successfully
either.

As a result, this commit:

* Adds testing our full matrix with Ruby 3.4
* Drops testing with Ruby 3.1
* Drops testing with Rails 7.1

I think this is a good balance of testing the latest versions of Ruby
and Rails while maintaining some backward compatibility in a manner that
doesn't require a degree in YAML to understand.