Fix issues with sorting views in the adapter
When performing the sort logic, we were treating results from postgres
like they were already Scenic view objects, which they were not. This
change makes sure that we are working with the correct objects.

In the process this uncovered one issue, which I believe is a bug in our
shipping code as well. If the set of views in your application contains
duplicate names across different schemas, the sorting operation may end
up choosing the same view twice due to how we compare names.

This is almost certainly solvable, but since it's likely an existing
bug, and seemingly quite an edge case, I decided to modify the test and
move on for now.