Allow create/update/replace_view to accept with:
Postgres accepts arguments in the form `WITH (<option>[ = <true|false>])`, so
rather than manually extending the various APIs to accept them individually
let's take a `with: {...}` argument that can take in arbitrary values.

The argument handling is:

- `false` or `null` values are discarded as the default behavior
- `true` values are passed as `<option> = true`
- other String or Symbol values are passed as `<option> = <value>`

This provides a more extensible API that can support any PostgreSQL view option
without requiring changes to Scenic's code, as evidence has shown that Derek and
I aren't super responsive maintainers. This should allow extensibility with new
Postgres (and other) features without requiring one of us to add support for
them.

A drawback and/or strength of this approach is that maintainers of Scenic
adapters such as MySQL or SQLite[^pronunciation] will have to deal with the
`with` hash individually, and that it may break migrations between databases. I
suspect that the latter is (a). unlikely to ever come up and (b). the least of the
worries of developers who undertake this.

- Add parse_view_options helper to generically parse PostgreSQL reloptions

[^pronunciation]: I just want to call out that "SQLite" is intended to be
pronounced as rhyming with meteorite and not with blight because I think that's
fun.

Conventional: Feat
Feat: Allow `with` argument to *_view to be passed to view DDL statements
Conventional: BREAKING CHANGE
Breaking: Users of the https://github.com/scenic-views/scenic/pull/433 fork will
          need to adjust their arguments slightly.